Release-ready: packaging, CI, docs + IPC/monitor fixes
Packaging / release
- PEP 517 + PEP 639 pyproject (build-system, dynamic version, project URLs,
  classifiers, package-data for the GUI stylesheet) and MANIFEST.in.
  `uv build` now produces sdist + wheel; the wheel installs cleanly and the
  throtl-cli/throtl-daemon/throtl-gui entry points resolve.
- Add CI workflow (ruff + unittest matrix 3.11-3.13, Xvfb GTK4 GUI tests,
  build), CHANGELOG.md, CONTRIBUTING.md and docs/RELEASING.md.
- README: badges, real clone URL, build/lint instructions.
- Makefile: lint / lint-fix / check / build / clean targets, ruff-aware.
- setup: install docs into /opt, drop stale __pycache__, remove PATH symlinks
  on uninstall.

Fixes
- protocol: buffer bytes after a message newline so coalesced responses/events
  and pipelined requests are no longer discarded (sporadically lost events).
- monitor: reap the nethogs subprocess and close its pipe on stop (zombie/leak).
- daemon: monitor_factory=None now genuinely disables monitoring.
- lint cleanup: unused imports/locals, typos, plus a reproducible ruff config.

+++ throtl/units.py
@@ -53,7 +53,7 @@ def parse_rate(value) -> int:
     if isinstance(value, (int, float)):
         if value < 0:
             raise ValueError(f"negative Rate ungueltig: {value!r}")
-        return int(round(value))
+        return round(value)
     if not isinstance(value, str):
         raise ValueError(f"ungueltige Rate: {value!r}")
     text = value.strip().lower().replace(" ", "")
@@ -68,7 +68,7 @@ def parse_rate(value) -> int:
                 break
             if amount < 0:
                 raise ValueError(f"negative Rate ungueltig: {value!r}")
-            return int(round(amount * factor))
+            return round(amount * factor)
     # nackte Zahl -> kbit/s
     try:
         amount = float(text)
@@ -76,7 +76,7 @@ def parse_rate(value) -> int:
         raise ValueError(f"ungueltige Rate: {value!r}") from None
     if amount < 0:
         raise ValueError(f"negative Rate ungueltig: {value!r}")
-    return int(round(amount))
+    return round(amount)
 
 
 def format_rate(kbit_per_s, unit: str = "auto", precision: int = 1) -> str:
@@ -151,7 +151,7 @@ def parse_rate_in_unit(value, unit: str = "kbps"):
     if value is None:
         return None
     if isinstance(value, (int, float)):
-        return int(round(float(value) * _UNIT_TO_KBIT.get(unit, 1.0)))
+        return round(float(value) * _UNIT_TO_KBIT.get(unit, 1.0))
     text = (value or "").strip().replace(",", ".")
     if not text:
         return None
@@ -166,7 +166,7 @@ def parse_rate_in_unit(value, unit: str = "kbps"):
         raise ValueError(f"ungueltige Rate: {value!r}") from None
     if number < 0:
         raise ValueError(f"negative Rate ungueltig: {value!r}")
-    return int(round(number * _UNIT_TO_KBIT.get(unit, 1.0)))
+    return round(number * _UNIT_TO_KBIT.get(unit, 1.0))
 
 
 def format_rate_for_entry(kbit_per_s, unit: str = "kbps", precision: int = 3) -> str:
